Added a new property in class Robot, called "baseHp".
This allows the player robot to be restored to full capacity between fights, via the method "repair".

In class CombatScene, a method called "checkBots" has been added. It checks if the player has

been defeated or if the five enemy robots have been destroyed.

If the enemy is destroyed, it is removed from the SKScene, and a new random robot takes its place.